python/requirements.txt
numpy>=1.11
pandas>=0.18
scikit-learn>=0.18
scipy>=0.18
pillow>=2.0.0